+++ api/typescript/src/engine/lib/kurtosis_context/grpc_web_engine_client.ts
@@ -11,8 +11,11 @@ import type {
     DestroyEnclaveArgs,
     GetEnclavesResponse,
     GetEngineInfoResponse,
-    StopEnclaveArgs
+    StopEnclaveArgs,
+    GetUserServiceLogsArgs,
+    GetUserServiceLogsResponse
 } from "../../kurtosis_engine_rpc_api_bindings/engine_service_pb";
+import {NO_ERROR_ENCOUNTERED_BUT_RESPONSE_FALSY_MSG} from "../consts";
 
 export class GrpcWebEngineClient implements GenericEngineClient {
     private readonly client: EngineServiceClientWeb
@@ -28,7 +31,7 @@ export class GrpcWebEngineClient implements GenericEngineClient {
             this.client.getEngineInfo(emptyArg, {}, (error: grpc_web.RpcError | null, response?: GetEngineInfoResponse) => {
                 if (error === null) {
                     if (!response) {
-                        resolve(err(new Error("No error was encountered but the response was still falsy; this should never " + "happen")));
+                        resolve(err(new Error(NO_ERROR_ENCOUNTERED_BUT_RESPONSE_FALSY_MSG)));
                     } else {
                         resolve(ok(response!));
                     }
@@ -56,7 +59,7 @@ export class GrpcWebEngineClient implements GenericEngineClient {
             this.client.createEnclave(createEnclaveArgs, {}, (error: grpc_web.RpcError | null, response?: CreateEnclaveResponse) => {
                 if (error === null) {
                     if (!response) {
-                        resolve(err(new Error("No error was encountered but the response was still falsy; this should never happen")));
+                        resolve(err(new Error(NO_ERROR_ENCOUNTERED_BUT_RESPONSE_FALSY_MSG)));
                     } else {
                         resolve(ok(response!));
                     }
@@ -118,8 +121,7 @@ export class GrpcWebEngineClient implements GenericEngineClient {
             this.client.clean(cleanArgs, {}, (error: grpc_web.RpcError | null, response?: CleanResponse) => {
                 if (error === null) {
                     if (!response) {
-                        resolve(err(new Error("No error was encountered but the response was still falsy; this " +
-                            "should never happen")));
+                        resolve(err(new Error(NO_ERROR_ENCOUNTERED_BUT_RESPONSE_FALSY_MSG)));
                     } else {
                         resolve(ok(response!));
                     }
@@ -144,7 +146,7 @@ export class GrpcWebEngineClient implements GenericEngineClient {
             this.client.getEnclaves(emptyArg, {}, (error: grpc_web.RpcError | null, response?: GetEnclavesResponse) => {
                 if (error === null) {
                     if (!response) {
-                        resolve(err(new Error("No error was encountered but the response was still falsy; this should never happen")));
+                        resolve(err(new Error(NO_ERROR_ENCOUNTERED_BUT_RESPONSE_FALSY_MSG)));
                     } else {
                         resolve(ok(response!));
                     }
@@ -160,5 +162,28 @@ export class GrpcWebEngineClient implements GenericEngineClient {
         }
 
         return ok(getEnclavesResponseResult.value);
+    }
+
+    public async getUserServiceLogs(getUserServiceLogsArgs: GetUserServiceLogsArgs): Promise<Result<GetUserServiceLogsResponse, Error>> {
+        const getUserServiceLogsPromise: Promise<Result<GetUserServiceLogsResponse, Error>> = new Promise((resolve, _unusedReject) => {
+            this.client.getUserServiceLogs(getUserServiceLogsArgs, {},(error: grpc_web.RpcError  | null, response?: GetUserServiceLogsResponse) => {
+                if (error === null) {
+                    if (!response) {
+                        resolve(err(new Error(NO_ERROR_ENCOUNTERED_BUT_RESPONSE_FALSY_MSG)))
+                    } else {
+                        resolve(ok(response));
+                    }
+                } else {
+                    resolve(err(error))
+                }
+            })
+        })
+
+        const getUserServiceLogsResponseResult: Result<GetUserServiceLogsResponse, Error> = await getUserServiceLogsPromise;
+        if (getUserServiceLogsResponseResult.isErr()) {
+            return err(getUserServiceLogsResponseResult.error);
+        }
+
+        return ok(getUserServiceLogsResponseResult.value);
     }
 }