+++ docs/changelog.md
@@ -1,4 +1,8 @@
 # TBD
+### Features
+* Added `KurtosisContext.GetUserServiceLogs` method in `golang` and `typescript` api libraries
+* Added the public documentation for the new `KurtosisContext.GetUserServiceLogs` method
+* Added `GetUserServiceLogs` in Kurtosis engine gateway
 
 ### Changes
 * Add back old enclave continuity test
@@ -7,7 +11,6 @@
 * * Implemented IP address references for services
 
 # 0.50.2
-
 ### Fixes
 * Fixes how the push cli artifacts & publish engine runs by generating kurtosis_version before hand